test: 테스트 로직, 함수명 수정
